Add NumberInput in place of time and duration inputs
PlayControls handles the ranges for the time and duration values:
e.g. time's minimum value is 0, but its maximum cannot be larger than duration.
If duration is set to any lower value than time, time is set to the new duration.
Also adding duration state for storing the duration changes.