Synchronized scrolling
When scrolling the ruler, it should scroll keyframe list in sync and vice versa (horizontally).
When scrolling the tracklist, it should scroll keyframe list in sync and vice versa (vertically).

Creating a context to share refs of the scrollable elements, then setting scrollLeft or scrollTop of the ref
can be used for synchronizing the scrolls.